+++ mirtop/importer/srnabench.py
@@ -162,6 +162,7 @@ def _read_iso(fn):
 
 
 def _translate(isomirs, description):
+    # TODO: make unit test for this
     iso = []
     labels = isomirs.split("@")
     logger.debug("TRANSLATE::label:%s" % isomirs)
@@ -177,6 +178,10 @@ def _translate(isomirs, description):
         if label.find("lv3p") > -1:
             iso.append("iso_3p:%s" % number_nts)
         if label.find("lv5p") > -1:
+            if number_nts.startswith("+"):
+                number_nts = number_nts.replace("+", "-")
+            else:
+                number_nts = number_nts.replace("-", "+")
             iso.append("iso_5p:%s" % number_nts)
         if label.find("nta") > -1:
             number_nts = label.split("|")[1].split("#")[-1]
@@ -196,13 +201,13 @@ def _translate(isomirs, description):
 def _iso_snp(pos):
     iso = []
     if pos > 1 and pos < 8:
-        iso.append("iso_snp_seed")
+        iso.append("iso_snv_seed")
     elif pos == 8:
-        iso.append("iso_snp_central_offset")
+        iso.append("iso_snv_central_offset")
     elif pos > 8 and pos < 13:
-        iso.append("iso_snp+central")
+        iso.append("iso_snv_central")
     elif pos > 12 and pos < 18:
-        iso.append("iso_snp_central_supp")
+        iso.append("iso_snv_central_supp")
     else:
-        iso.append("iso_snp")
+        iso.append("iso_snv")
     return iso